Fixes issue related to multiple views into a file
There seems to be a bug in the Sublime Text 3 API that prevents
on_selection_modified from working correctly with multiple views
into the same file. The command always receives thI original view
associated with the file. I updated the command with a workaround
that grabs the active view of the current window before doing
anything with it.